Revert "5 criar função relatório"
