--- .vscode/settings.json
@@ -1,9 +1,0 @@
-{
-    "files.associations": {
-        "deque": "cpp",
-        "string": "cpp",
-        "vector": "cpp",
-        "iostream": "cpp",
-        "limits": "cpp"
-    }
-}
